+++ source/main/zettlr-window.js
@@ -84,7 +84,8 @@ class ZettlrWindow {
       webPreferences: {
         // Zettlr needs all the node features, so in preparation for Electron
         // 5.0 we'll need to explicitly request it.
-        nodeIntegration: true
+        nodeIntegration: true,
+        enableRemoteModule: false
       },
       backgroundColor: '#fff',
       scrollBounce: true, // The nice scrolling effect for macOS